add local variables

+++ src/main/java/desu/fullStack/lang/ast/ConstantExpression.java
@@ -1,13 +1,13 @@
 package desu.fullStack.lang.ast;
 
 public class ConstantExpression implements Expression {
-	private int value;
+	private float value;
 
-	public ConstantExpression(int value) {
+	public ConstantExpression(float value) {
 		this.value = value;
 	}
 	
-	public int getValue() {
+	public float getValue() {
 		return value;
 	}
